+++ src/main/java/fr/mkadia/mkadiaapi/repositories/OrderRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrderRepository extends JpaRepository<Order, Integer>, JpaSpecificationExecutor<Order> {
+public interface OrderRepository extends JpaRepository<Order, Long>, JpaSpecificationExecutor<Order> {
     Page<Order> findAll(Specification<Order> spec, Pageable pageable);
     Page<Order> findAllByUser(User user, Pageable pageable);
 }
