+++ src/main/resources/db/migration/V1_14__order_items_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Order_Items (
-    order_item_id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-    order_id INT NOT NULL,
-    product_id INT NOT NULL,
+    order_item_id BIGINT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+    order_id BIGINT NOT NULL,
+    product_id BIGINT NOT NULL,
     quantity INT NOT NULL CHECK (quantity > 0),
     price DECIMAL(10, 2) NOT NULL CHECK (price >= 0),
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
